* sorted intra-package imports (check out the funky relative import statement) * added (the start of) a sample() method to pyGP.GP - not sure if it makes sense yet
Note that with packages, the `if __name__ == "__main__"` idiom starts to make less and less sense. Also, with relative imports, it's impossible to run a `package` file as a script so we can do EITHER intra-package imports OR use if name == main stuff. Hence the adding of a testing directory...